Upgrade to Laravel 5.8

+++ database/migrations/2018_08_03_210351_create_task_lists_table.php
@@ -14,8 +14,8 @@ class CreateTaskListsTable extends Migration
     public function up()
     {
         Schema::create('task_lists', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigIncrements('id');
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->string('name');
             $table->timestamps();
